refactor(solver): split narrowing core helpers
